Fix review findings.
